Object Mapper – Revert refactoring, fix missing items in updateObjects response from Sync API
This is because QueryBuilder is passed by reference, causing later query parts (and importantly, conflicting query parts) to be joined with earlier ones, if the same QueryBuilder instance is passed to several methods that add to the query. In other words, each QueryBuilder seems to hold a global query that is manipulated even if the instance is passed as parameter to different methods, which was thought to duplicate the instance.